Introduce the way to mark views for binding doc pages to
Use DocViewMixin to mark a view which a doc page can be bound
to. Login page is processed in its own way, because its view is inside
openstack_auth lib.

To determine if a 'Click Me' doc page button should be rendered for a
view, when rendering its contents, check if its URL already exists in
one the DocPage model objects stored in the database (see next
commit).

Change-Id: I7547e3fbab4f5dcf31b4ffe67fe61b9d8d09c374